+++ sonarr.py
@@ -8,6 +8,9 @@ load_dotenv()
 API_KEY = os.getenv("SONARR_API_KEY")
 BASE_URL = os.getenv("SONARR_BASE_URL")
 
+DRY_RUN = os.getenv("DRY_RUN", "False").lower() in ("true", "1", "t")
+
+
 def get_sonarr_id(tvdb_id):
     """
     Retrieves the Sonarr ID, title, and size on disk for a TV series with the given TVDB ID.
@@ -30,10 +33,11 @@ def get_sonarr_id(tvdb_id):
         series_data = response.json()
         if not series_data:
             return None, None, None
-        return series_data[0]['id'], series_data[0]['title'], series_data[0]['statistics']['sizeOnDisk']
+        return series_data[0]["id"], series_data[0]["title"], series_data[0]["statistics"]["sizeOnDisk"]
     else:
         raise Exception(f"Fetching Sonarr ID failed with status code {response.status_code}")
-    
+
+
 def find_and_delete_series(tvdb_id):
     """
     Finds a TV series in Sonarr by its TVDB ID and deletes it along with its files.
@@ -49,11 +53,12 @@ def find_and_delete_series(tvdb_id):
     """
     sonarr_id, title, size_on_disk = get_sonarr_id(tvdb_id)
     if sonarr_id is None or title is None:
-        raise Exception('Fetching Sonarr ID failed')
-    
+        raise Exception("Fetching Sonarr ID failed")
+
     delete_unplayed_series(sonarr_id, title, size_on_disk)
 
     return size_on_disk
+
 
 def delete_unplayed_series(sonarr_id, title, size_on_disk):
     """
@@ -69,10 +74,14 @@ def delete_unplayed_series(sonarr_id, title, size_on_disk):
     """
     url = f"{BASE_URL}/series/{sonarr_id}"
     headers = {"X-Api-Key": API_KEY}
-    params = {'deleteFiles': 'true'}
+    params = {"deleteFiles": "true"}
+
+    if DRY_RUN:
+        print(f"SONARR :: DRY RUN :: {title} would be deleted. {convert_bytes(size_on_disk)} would be freed up")
+        return
 
     response = requests.delete(url, headers=headers, params=params)
     if response.status_code == 200:
-        print(f"{title} deleted successfully. {convert_bytes(size_on_disk)} freed up.")
+        print(f"SONARR :: {title} deleted successfully. {convert_bytes(size_on_disk)} freed up")
     else:
         raise Exception(f"Deletion failed with status code {response.status_code}")
